moved classes to models folder
